feat(ui): add simple Backdrop component for modals
