+++ packages/ui/src/design-system/backdrop/index.ts
@@ -1,0 +1,2 @@
+export { Backdrop } from './backdrop.component';
+export type { BackdropProps } from './backdrop.component';
