+++ packages/ui/src/design-system/dialog/dialog-root.css.ts
@@ -1,10 +1,5 @@
 import { style } from '../../design-tokens';
 
-export const dialogOverlay = style({
-  position: 'fixed',
-  inset: 0,
-});
-
 export const dialogContent = style({
   position: 'fixed',
   top: '50%',
