+++ packages/ui/src/design-system/side-drawer/side-drawer-content.component.tsx
@@ -3,9 +3,10 @@ import React from 'react';
 import * as Dialog from '@radix-ui/react-dialog';
 import classNames from 'classnames';
 
+import { Backdrop } from '../backdrop';
+
 import { ContentCard } from './side-drawer-content-card.component';
 import * as cx from './side-drawer-content.css';
-import { Overlay } from './side-drawer-overlay.component';
 
 // TODO refactor: pass z-indices config on app level: https://vanilla-extract.style/documentation/packages/dynamic/
 type Props = Dialog.DialogPortalProps &
@@ -18,7 +19,9 @@ export const Content = ({
   ...props
 }: Readonly<Props>): JSX.Element => (
   <Dialog.Portal container={container}>
-    <Overlay zIndex={zIndex} />
+    <Dialog.Overlay asChild>
+      <Backdrop zIndex={zIndex} />
+    </Dialog.Overlay>
     <Dialog.Content
       {...props}
       className={classNames(cx.container)}
